// src/app/api/izdelki/model/izdelekDTO.ts
/**
 * Katalog izdelkov
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LastnostDTO } from './lastnostDTO';
import { SlikaDTO } from './slikaDTO';

export interface IzdelekDTO {
  id_izdelek?: number;
  naziv?: string;
  opis?: string;
  cena?: number;
  aktiven?: boolean;
  datum_dodajanja?: string;
  datum_spremembe?: string;
  zaloga?: number;
  slike?: Array<SlikaDTO>;
  slikeDodaj?: Array<SlikaDTO>;
  slikeBrisi?: Array<SlikaDTO>;
  lastnosti?: Array<LastnostDTO>;
  lastnostiDodaj?: Array<LastnostDTO>;
  lastnostiBrisi?: Array<LastnostDTO>;
}